requirements.txt
Flask
gunicorn
wtforms
pymysql
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Jinja2
Flask-WTF
cryptography
psycopg2
psycopg2-binary
email_validator
Flask-Mail







